backend/hotel_mcp_agent/pyproject.toml
[project]
name = "hotel-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "mcp>=1.13.1",
    "openai>=1.101.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
